+++ bench_tps/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+authors = ["Solana Maintainers <maintainers@solana.com>"]
+edition = "2018"
+name = "bench_tps"
+version = "0.1.0"
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "http://solana.com/"
+
+[dependencies]
+clap = "2.32.0"
+rayon = "1.0.3"
+
+[dependencies.solana]
+path = ".."
+version = "0.11.0"
+
+[dependencies.solana-sdk]
+path = "../sdk"
+version = "0.11.0"
+
+[dependencies.solana-metrics]
+path = "../metrics"
+version = "0.11.0"
+
+[dependencies.solana-drone]
+path = "../drone"
+version = "0.11.0"
